+++ src/App.css
@@ -1,81 +1,49 @@
 #root {
-  height: 100%;
-  width: 100%;
-  padding: 1rem;
-  text-align: left;
-  box-sizing: border-box
-}
-
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.react:hover {
-  filter: drop-shadow(0 0 2em #61dafbaa);
-}
-
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
+    height: 100%;
+    width: 100%;
+    padding: 1rem;
+    text-align: left;
+    box-sizing: border-box
   }
-  to {
-    transform: rotate(360deg);
+  
+  .logo {
+    height: 6em;
+    padding: 1.5em;
+    will-change: filter;
+    transition: filter 300ms;
   }
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  a:nth-of-type(2) .logo {
-    animation: logo-spin infinite 20s linear;
+  .logo:hover {
+    filter: drop-shadow(0 0 2em #646cffaa);
   }
-}
-
-.card {
-  padding: 2em;
-}
-
-.read-the-docs {
-  color: #888;
-}
-
-h1 {
-  margin-bottom: 1rem;
-}
-
-.fc { /* the calendar root */
-  max-width: 1280px;
-  margin: 0 auto;
-}
-
-.app-container {
-  display: flex;
-  min-height: 100vh;
-}
-
-.sidebar {
-  width: 300px;
-  padding: 2rem;
-  border-right: 1px solid #ddd;
-}
-
-.calendar-container {
-  flex-grow: 1;
-  padding: 2rem;
-}
-
-.sidebar-section {
-  margin-bottom: 2rem;
-}
-
-.sidebar-section ul {
-  margin: 0;
-  padding-left: 1.5rem;
-}
-
-.sidebar-section li {
-  margin: 0.5rem 0;
-}
+  .logo.react:hover {
+    filter: drop-shadow(0 0 2em #61dafbaa);
+  }
+  
+  @keyframes logo-spin {
+    from {
+      transform: rotate(0deg);
+    }
+    to {
+      transform: rotate(360deg);
+    }
+  }
+  
+  @media (prefers-reduced-motion: no-preference) {
+    a:nth-of-type(2) .logo {
+      animation: logo-spin infinite 20s linear;
+    }
+  }
+  
+  .card {
+    padding: 2em;
+  }
+  
+  .read-the-docs {
+    color: #888;
+  }
+  
+  h1 {
+    margin-bottom: 1rem;
+  }
+  
+
