+++ src/index.css
@@ -66,3 +66,10 @@ button:focus-visible {
     background-color: #f9f9f9;
   }
 }
+
+/* This root refers to clerk's log in page */
+#root {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
